requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
mysqlclient
requests
python-dotenv

# Optional for JWT-protected API:
PyJWT